+++ docs/changelog.rst
@@ -5,6 +5,7 @@ v1.1.0 - not yet released
 ------------------------------
 
 * 293: Add order by last updated to order options - http://github.com/mPowering/django-orb/issues/issue/293
+* 228: Allow ordering by rating - http://github.com/mpowering/django-orb/issues/issue/228
 
 v1.0.1
 ---------------------